Added base structure for GUI support: CControlManagerUI, base CControlUI and simple CButtonUI

+++ src/event.cpp
@@ -1,6 +1,6 @@
 #include "../include/event.h"
 
-CEvent::CEvent(EEventController controller, uint32 id, double x, double y) {
+CEvent::CEvent(EEventController controller, uint32 id, int32 x, int32 y) {
 	m_controller = controller;
 	m_id = id;
 	m_x = x;
